modal: refactor sort seed loader
